lllllll

+++ app/__init__.py
@@ -141,16 +141,17 @@ def show_one_thing(id):
 def add_a_thing():
     # Get the data from the form
     name  = request.form.get("name")
-    price = request.form.get("price")
 
     # Sanitise the inputs
     name = html.escape(name)
-    price = html.escape(price)
+
+    #Session UID
+    uid = session["user_id"]
 
     with connect_db() as client:
         # Add the thing to the DB
-        sql = "INSERT INTO things (name, price) VALUES (?, ?)"
-        values = [name, price]
+        sql = "INSERT INTO things (name, uid) VALUES (?, ?)"
+        values = [name, uid]
         client.execute(sql, values)
 
         # Go back to the home page
@@ -193,13 +194,21 @@ def add_a_user():
 @app.get("/delete/<int:id>")
 def delete_a_thing(id):
     with connect_db() as client:
+        user_id = session["user_id"]
         # Delete the thing from the DB
-        sql = "DELETE FROM things WHERE id=?"
-        values = [id]
+        sql = "DELETE FROM things WHERE id=? and user_id = ?"
+        values = [id, user_id]
         client.execute(sql, values)
 
         # Go back to the home page
         flash("Thing deleted", "warning")
         return redirect("/things")
+    
+@app.get("/logout")
+def logout():
+    session.pop("user_id")
+    session.pop("user_name")
+    flash("Logout", "error")
+    return redirect("/")
